% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sageParenting.R
\docType{data}
\name{exclamation}
\alias{exclamation}
\title{Exclamations}
\format{
An object of class \code{character} of length 16.
}
\usage{
exclamation
}
\description{
Collection of phrases used by parents to start a comment.
  Not exported.
}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
}
\keyword{datasets}
