% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sageParenting.R
\docType{data}
\name{frown}
\alias{frown}
\title{Text-based Frowns}
\format{
An object of class \code{character} of length 28.
}
\usage{
frown
}
\description{
Collection of text-based frowns to show additional displeasure.
  Based on emojis found in the [List of emoticons on Wikipedia]
  (https://en.wikipedia.org/wiki/List_of_emoticons)

  Not exported.
}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
}
\keyword{datasets}
