% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sageParenting.R
\docType{data}
\name{phrase}
\alias{phrase}
\title{Phrases}
\format{
An object of class \code{character} of length 62.
}
\usage{
phrase
}
\description{
Collection of phrases muttered by parents.
  Not exported.
}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
}
\keyword{datasets}
