% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sageParenting.R
\name{sage.parenting}
\alias{sage.parenting}
\title{Sage Parenting}
\usage{
sage.parenting()
}
\value{
`string` with all the feels
}
\description{
Just like your parents, only in your R session.
}
\details{
Are you a child of the 1980s or 1990s? Where your parents the
  masters of crushing your soul and spirit? Do you miss their support? Ever
  feel too good? Like you are living your best life? Fire off the command
  `sage.parenting()` to be put back in your place. Nothing like
  a comment from mom or dad to adjust your attitude.
}
\examples{
set.seed(13)
sage.parenting()
# [1] "JESUS CHRIST! If you don't stop acting like this, we will go home right now. >:["

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
}
